+++ packages/web/src/types.ts
@@ -1,4 +1,3 @@
-import { Org } from "@sourcebot/db";
 import { z } from "zod";
 
 export const orgMetadataSchema = z.object({
@@ -6,8 +5,3 @@ export const orgMetadataSchema = z.object({
 })
 
 export type OrgMetadata = z.infer<typeof orgMetadataSchema>;
-
-export const getOrgMetadata = (org: Org): OrgMetadata | null => {
-    const currentMetadata = orgMetadataSchema.safeParse(org.metadata);
-    return currentMetadata.success ? currentMetadata.data : null;
-}